requirements.txt
Package                   Version
------------------------- -----------
amqp                      5.2.0
asgiref                   3.8.1
asttokens                 2.4.1
attrs                     23.2.0
beautifulsoup4            4.12.3
billiard                  4.2.0
celery                    5.4.0
certifi                   2024.2.2
charset-normalizer        3.3.2
click                     8.1.7
click-didyoumean          0.3.1
click-plugins             1.1.1
click-repl                0.3.0
contourpy                 1.2.1
contrib                   0.3.0
cycler                    0.12.1
decorator                 5.1.1
Django                    4.2.13
django-apps               0.1.0
django-bootstrap-v5       1.0.11
django-celery-results     2.5.1
django-core               1.4.1
django-crontab            0.7.1
django_db                 0.0.7
django-environ            0.11.2
django-http-method        1.2.1
django-shortcuts          1.6
django_urls               1.1.3
environ                   1.0
executing                 2.0.1
fonttools                 4.52.1
googlemaps                4.10.0
idna                      3.7
ipython                   8.25.0
jedi                      0.19.1
jsonschema                4.22.0
jsonschema-specifications 2023.12.1
kiwisolver                1.4.5
kombu                     5.3.7
matplotlib                3.9.0
matplotlib-inline         0.1.7
numpy                     1.26.4
optional-django           0.3.0
packaging                 24.0
pandas                    2.2.2
parso                     0.8.4
pathlib                   1.0.1
pexpect                   4.9.0
phonenumbers              8.13.37
pillow                    10.3.0
pip                       24.0
prompt_toolkit            3.0.47
psycopg2-binary           2.9.9
ptyprocess                0.7.0
pure-eval                 0.2.2
Pygments                  2.18.0
PyMySQL                   1.1.1
pyparsing                 3.1.2
python-dateutil           2.9.0.post0
pytz                      2024.1
PyYAML                    6.0.1
react                     4.3.0
referencing               0.35.1
requests                  2.32.2
rpds-py                   0.18.1
setuptools                69.5.1
shortcuts                 0.11.0
six                       1.16.0
soupsieve                 2.5
sqlparse                  0.5.0
stack-data                0.6.3
toml                      0.10.2
traitlets                 5.14.3
tzdata                    2024.1
urllib3                   2.2.1
utils                     1.0.2
vine                      5.1.0
wcwidth                   0.2.13
wheel                     0.43.0